+++ src/main/java/nextstep/subway/dto/StationResponse.java
@@ -1,6 +1,6 @@
 package nextstep.subway.dto;
 
-import nextstep.subway.domain.Station;
+import nextstep.subway.domain.station.Station;
 
 import java.time.LocalDateTime;
